Push hacky logic into sbs package
